requirement.txt
asgiref==3.8.1
Django==5.1.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-filter==0.2.2
mysql-connector==2.2.9
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
